+++ Mods ChoGGi/Click To Move Camera/metadata.lua
@@ -8,14 +8,13 @@ return PlaceObj("ModDef", {
 		}),
 	},
 	"title", "Click To Move Camera",
-	"version", 3,
-	"version_major", 0,
-	"version_minor", 3,
-
-	"image", "Preview.png",
 	"id", "ChoGGi_ShiftClickToMove",
 	"steam_id", "1590473613",
 	"pops_any_uuid", "66775804-4873-47bf-8ffb-005195ebb5c8",
+	"version", 3,
+	"version_major", 0,
+	"version_minor", 3,
+	"image", "Preview.png",
 	"author", "ChoGGi",
 	"lua_revision", 249143,
 	"code", {