+++ Mods ChoGGi/Construction Show Hex Buildable Grid/metadata.lua
@@ -21,22 +21,22 @@ return PlaceObj("ModDef", {
 		"Code/Script.lua",
 	},
 	"has_options", true,
-	"description", [[Show different colours for different types of hexes during construction.
+	"description", [[Show different colours for different types of hexes during construction (disable in mod options).
 Press Numpad 0 to toggle grid anytime (rebind in game options).
 
 Passable (drones/rovers can drive here), Buildable (buildings can be placed here)
 Green = pass/build
 Yellow = no pass/build
-Blue = pass/no build (shows blue for pipes, but you can build some stuff under them)
+Blue = pass/no build
 Red = no pass/no build
-The colours use the centre of the hex, so they aren't exact.
-
-[url=https://steamcommunity.com/sharedfiles/filedetails/?id=1479851929]Show Hex Grid[/url]
+Shows blue for pipes, but you can build some stuff under them. The colours use the centre of the hex, so they could be off?
 
 
 Mod Options:
 Show during construction: If you don't want grids showing up during construction placement.
 Grid Size: Set the size of the grid area.
 Grid Opacity: Set opacity of grid icons.
+
+[url=https://steamcommunity.com/sharedfiles/filedetails/?id=1479851929]Show Hex Grid[/url]
 ]],
 })