+++ Mods ChoGGi/DroneHub Visual Drone Load/metadata.lua
@@ -1,8 +1,8 @@
 return PlaceObj("ModDef", {
 	"title", "DroneHub Visual Drone Load",
-	"version", 5,
+	"version", 6,
 	"version_major", 0,
-	"version_minor", 5,
+	"version_minor", 6,
 	"image", "Preview.png",
 	"id", "ChoGGi_DroneHubVisualDroneLoad",
 	"steam_id", "1757808500",