+++ Mods ChoGGi/Drones Harvest Rocks/Code/Script.lua
@@ -5,9 +5,11 @@ local IsValid = IsValid
 local Sleep = Sleep
 local PlayFX = PlayFX
 local GetRandomPassableAround = GetRandomPassableAround
+local TransformToStockpile = WasteRockObstructor.TransformToStockpile
+local efSelectable = const.efSelectable
 
-function Drone:ChoGGi_RockRemove(obj)
-	self:Face(obj, 100)
+function Drone:ChoGGi_RockRemove(rock)
+	self:Face(rock, 100)
 
 	PlayFX("Deconstruct", "start", self)
 	self:PlayState("gatherStart")
@@ -20,51 +22,66 @@ function Drone:ChoGGi_RockRemove(obj)
 	while scale > 0 do
 		Sleep(dur)
 		scale = scale - 15
-		obj:SetScale(scale)
+		rock:SetScale(scale)
 		Sleep(dur)
 	end
 
 	self:PlayState("gatherEnd")
 	PlayFX("Deconstruct", "end", self)
 
-	obj:TransformToStockpile(self)
-end
-
-local efSelectable = const.efSelectable
-local function DroneRemoveRock(obj)
-	if not obj.TransformToStockpile then
-		return
-	end
-	local rock_pos = obj:GetPos()
-
-	local drone = MapFindNearest(rock_pos, "map", "Drone", efSelectable, function(d)
-		if d.command == "Idle" then
-			return true
-		end
-	end)
-	if not drone then
-		return
-	end
-
-	local pos = GetRandomPassableAround(rock_pos, 500)
-	if pos then
-		GetCommandFunc(drone)(drone, "Goto", pos:SetTerrainZ())
-
-		-- good enough
-		while drone.command ~= "Idle" and drone.command ~= "GoHome" do
-			Sleep(250)
-		end
-
-		GetCommandFunc(drone)(drone, "ChoGGi_RockRemove", obj)
+	if rock.TransformToStockpile then
+		rock:TransformToStockpile(self)
+	else
+		TransformToStockpile(rock, self)
 	end
 end
 
--- If it's a waste rock and it doesn't DoesNotObstructConstruction then it's "safe" (read: non-cheaty) to remove
+local function DroneRemoveRock(rock)
+--~ 	if not rock.TransformToStockpile then
+--~ 		return
+--~ 	end
+	if rock:IsKindOf("WasteRockObstructor") or rock:IsKindOf("StoneSmall") then
+		local rock_pos = rock:GetPos()
+
+		local drone = MapFindNearest(rock_pos, "map", "Drone", efSelectable, function(d)
+			if d.command == "Idle" then
+				return true
+			end
+		end)
+		if not drone then
+			return
+		end
+
+		local pos = GetRandomPassableAround(rock_pos, 500)
+		if pos then
+
+			-- only add one
+			if not rock:GetAttaches("RotatyThing") then
+				-- let user know
+				local rotate = PlaceObject("RotatyThing")
+				rock:Attach(rotate)
+--~ 				rock:SetAttachOffset(point(0, 0, rock:GetPos():SetTerrainZ(5000):z()))
+			end
+
+			GetCommandFunc(drone)(drone, "Goto", pos:SetTerrainZ())
+
+			-- good enough
+			while drone.command ~= "Idle" and drone.command ~= "GoHome" do
+				Sleep(250)
+			end
+
+			GetCommandFunc(drone)(drone, "ChoGGi_RockRemove", rock)
+		end
+	end
+end
+
 local orig_CanDemolish = DemolishModeDialog.CanDemolish
 function DemolishModeDialog:CanDemolish(pt, obj, ...)
 	obj = obj or SelectionMouseObj() or GetTerrainCursorObj()
-	if IsValid(obj) and obj:IsKindOf("DoesNotObstructConstruction") then
-		return true
+	if IsValid(obj) then
+		if obj:IsKindOf("WasteRockObstructor") or obj:IsKindOf("StoneSmall") then
+			return true
+		end
 	end
 
 	return orig_CanDemolish(self, pt, obj, ...)
@@ -74,9 +91,11 @@ local orig_OnMouseButtonDown = DemolishModeDialog.OnMouseButtonDown
 function DemolishModeDialog:OnMouseButtonDown(pt, button, obj, ...)
 	if button == "L" then
 		obj = obj or SelectionMouseObj() or GetTerrainCursorObj()
-		if IsValid(obj) and obj:IsKindOf("DoesNotObstructConstruction") then
-			CreateGameTimeThread(DroneRemoveRock, obj)
-			return "break"
+		if IsValid(obj) then
+			if obj:IsKindOf("WasteRockObstructor") or obj:IsKindOf("StoneSmall") then
+				CreateGameTimeThread(DroneRemoveRock, obj)
+				return "break"
+			end
 		end
 	end