+++ Mods ChoGGi/Logos/CommieMarsLogos/metadata.lua
@@ -25,6 +25,6 @@ I probably should've kept a list of credits... I didn't make any of these images
 You can use my [url=https://steamcommunity.com/sharedfiles/filedetails/?id=1815122803]Change Logo[/url] mod to change the logo of an existing game.
 
 This is for you (you know who you are):
-Commies are murders, so are Nazies (and Capitalists and Christans and Muslims and so on).
+Commies are murderers, so are Nazies (and Capitalists and Christans and Muslims and so on).
 Congratulations! You win an Interwebs!]],
 })