+++ Mods ChoGGi/Unit Thoughts/metadata.lua
@@ -12,9 +12,9 @@ return PlaceObj("ModDef", {
 	"lua_revision", 249143,
 	"steam_id", "2196814512",
 	"pops_any_uuid", "646cf5c2-b36e-40bd-b939-ab7b3fe73f06",
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.png",
 	"author", "ChoGGi",
 	"code", {
@@ -23,7 +23,7 @@ return PlaceObj("ModDef", {
 	"has_options", true,
 	"TagInterface", true,
 	"description", [[Select a unit (drone/rover/colonist/shuttle) to see what it's up to.
-Optionally show unit name and target
+Optionally show unit name, target, map grid target area.
 
 Mod Options:
 Enable Mod, Enable Text, Text Background, Text Opacity, Text Style